Split up some tests and add some delays
